correct cast is carried out in MOLogger.cpp
git-svn-id: https://svn.sourceforge.jp/svnroot/molds/MolDS/trunk@399 1136aad2-a195-0410-b898-f5ea1d11b9d8

+++ src/base/MOLogger.cpp
@@ -84,7 +84,7 @@ void MOLogger::DrawMO(vector<int> moIndeces){
       // output cube file
       ofs << this->messageCubeHeaderComment1;
       ofs << this->messageCubeHeaderComment2;
-      sprintf(data,"\t%d\t%e\t%e\t%e\n", this->molecule->GetAtomVect()->size(),
+      sprintf(data,"\t%d\t%e\t%e\t%e\n", (int)this->molecule->GetAtomVect()->size(),
                                          origin[XAxis], 
                                          origin[YAxis], 
                                          origin[ZAxis]);
